feat(v2-pie-interaction): 饼图图例交互
饼图图例激活的时候，默认带上legend-active交互，避免外部移除影响此交互行为的失效，end 交互同legend-active
