+++ src/plots/pie/interaction/pie-statistic-action.ts
@@ -1,10 +1,7 @@
-import { registerAction, registerInteraction, Util } from '@antv/g2';
-import Element from '@antv/g2/lib/geometry/element';
 import { Action } from '@antv/g2/lib/interaction';
 import { ComponentOption } from '@antv/g2/lib/interface';
-import { getDelegationObject } from '@antv/g2/lib/interaction/action/util';
 import { each, get } from '@antv/util';
-import { getStatisticData } from './utils';
+import { getStatisticData } from '../utils';
 
 /**
  * Pie 中心文本事件的 Action
@@ -77,58 +74,3 @@ export class StatisticAction extends Action {
     view.render(true);
   }
 }
-
-registerAction('pie-statistic', StatisticAction);
-registerInteraction('pie-statistic-active', {
-  start: [{ trigger: 'element:mouseenter', action: 'pie-statistic:change' }],
-  end: [{ trigger: 'element:mouseleave', action: 'pie-statistic:reset' }],
-});
-
-/**
- * 饼图 图例激活 action
- */
-export class PieLegendAction extends Action {
-  /**
-   * 获取激活的图形元素
-   */
-  private getActiveElements(): Element[] {
-    const delegateObject = getDelegationObject(this.context);
-    if (delegateObject) {
-      const view = this.context.view;
-      const { component, item } = delegateObject;
-      const field = component.get('field');
-      if (field) {
-        const elements = view.geometries[0].elements;
-        return elements.filter((ele) => ele.getModel().data[field] === item.value);
-      }
-    }
-    return [];
-  }
-
-  public active() {
-    const { view } = this.context;
-    const elements = this.getActiveElements();
-    elements.forEach((element) => {
-      const coordinate = element.geometry.coordinate;
-      if (coordinate.isPolar && coordinate.isTransposed) {
-        const { startAngle, endAngle } = Util.getAngle(element.getModel(), coordinate);
-        const middleAngle = (startAngle + endAngle) / 2;
-        /** offset 偏移 */
-        const r = 7.5;
-        const x = r * Math.cos(middleAngle);
-        const y = r * Math.sin(middleAngle);
-        const matrix = Util.transform(null, [['t', x, y]]);
-        element.shape.setMatrix(matrix);
-      }
-    });
-    view.render(true);
-  }
-
-  public reset() {}
-}
-
-registerAction('pie-legend', PieLegendAction);
-registerInteraction('pie-legend-active', {
-  start: [{ trigger: 'legend-item:mouseenter', action: 'pie-legend:active' }],
-  end: [{ trigger: 'legend-item:mouseleave', action: 'pie-legend:reset' }],
-});
